parser/requirements.full.txt
adal==1.2.2
aiohttp==3.6.2
aioredis==1.3.1
ansible-runner==2.0.0.0a5
ansiconv==1.0.0
asciichartpy==1.5.25
asgiref==3.3.4
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.12.3
Automat==20.2.0
azure-common==1.1.25
azure-keyvault==1.1.0
azure-nspkg==3.0.2
cachetools==4.0.0
certifi==2021.5.30
cffi==1.14.0
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
Cython==0.29.22
daphne==2.4.1
defusedxml==0.6.0
dictdiffer==0.8.1
distro==1.5.0
Django==2.2.16
django-auth-ldap==2.1.0
django-cors-headers==3.7.0
django-crum==0.7.5
django-extensions==2.2.9
django-guid==2.2.0
django-jsonfield==1.2.0
django-oauth-toolkit==1.1.3
django-pglocks==1.0.4
django-polymorphic==2.1.2
django-qsstats-magic==1.1.0
django-radius==1.3.3
django-redis==4.5.0
django-solo==1.1.3
django-split-settings==1.0.0
django-taggit==1.2.0
djangorestframework==3.12.1
djangorestframework-yaml==1.0.3
docutils==0.16
future==0.16.0
gitdb==4.0.2
GitPython==3.1.7
google-auth==1.11.3
hiredis==1.0.1
hyperlink==20.0.1
idna==2.9
idna-ssl==1.1.0
importlib-metadata==4.5.0
importlib-resources==5.1.4
incremental==17.5.0
irc==18.0.0
isodate==0.6.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.0
jaraco.logging==3.0.0
jaraco.stream==3.0.0
jaraco.text==3.2.0
Jinja2==2.11.2
JSON-log-formatter==0.3.0
jsonschema==3.2.0
kubernetes==11.0.0
lockfile==0.12.2
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
more-itertools==8.2.0
msgpack==1.0.0
msrest==0.6.11
msrestazure==0.6.3
multidict==4.7.5
netaddr==0.7.19
oauthlib==3.1.0
openshift==0.11.0
pexpect==4.7.0
pkgconfig==1.5.1
prometheus-client==0.7.1
psutil==5.8.0
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygerduty==0.38.2
PyHamcrest==2.0.2
PyJWT==2.1.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrad==2.3
pyrsistent==0.15.7
python-daemon==2.2.4
python-dateutil==2.8.1
python-ldap==3.3.1
python-string-utils==1.0.0
python3-openid==3.1.0
python3-saml==1.9.0
pytz==2019.3
PyYAML==5.4.1
receptorctl==1.0.0.0a2
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
requirements-parser==0.2.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
schedule==0.6.0
service-identity==18.1.0
six==1.14.0
slackclient==1.1.2
smmap==3.0.1
social-auth-app-django==3.1.0
social-auth-core==3.3.1
sqlparse==0.3.1
tacacs-plus==1.0
tempora==2.1.0
twilio==6.37.0
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.10.0.0
urllib3==1.25.8
uWSGI==2.0.18
uwsgitop==0.11
websocket-client==0.57.0
xmlsec==1.3.3
yarl==1.4.2
zipp==3.4.1
zope.interface==5.0.0
